requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI/LLM
google-generativeai==0.3.2

# HTTP client for Supabase integration
httpx==0.25.2
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CORS middleware
python-multipart==0.0.6
